% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAlbicansR_function.R
\name{orfToName}
\alias{orfToName}
\title{orfToName Function}
\usage{
orfToName(orfList)
}
\arguments{
\item{orfList}{Object of class "vector"}
}
\description{
This function access a pre-processed local database to convert
orf19 values to gene names
}
\examples{
orfList <- c("orf19.610","orf19.4884","orf19.5992")
nameToOrf(orfList)
}
