% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAlbicansR_function.R
\name{runGOEnrichment}
\alias{runGOEnrichment}
\title{runGOEnrichment Function}
\usage{
runGOEnrichment(geneList = NULL, type = "P",
  server = "selenium.joshuawang.com", port = 4445)
}
\arguments{
\item{geneList}{Object of class "vector". Should contain list of desired genes to be searched.}

\item{type}{Object of class "character". Type of ontology: P for Process, F for Function, C for Component. Defaults to P.}

\item{server}{Object of class "character". Users are encouraged to setup a local RSelenium instance. Defaults to a remote server.}

\item{port}{Object of class "integer". Defaults to port 4445.}
}
\description{
This function uses an RSelenium wrapper to access the CGD
GO Term Finder site, run the search, and download results with
corrected p-values.
}
\examples{
geneList <- c("orf19.610","orf19.4884","orf19.5992","WOR3","AHR1","CZF1")
runGOEnrichment(geneList,type='F')
}
